requirements.txt
streamlit
pandas
sqlalchemy
pymysql
redis
rq
python-dotenv
openai
numpy
google-generativeai
anthropic
